requirements.txt
feedparser==6.0.11
requests==2.31.0
google-cloud-aiplatform>=1.60.0
beautifulsoup4==4.12.3
python-dateutil==2.9.0
langdetect==1.0.9
